requirements.txt
requests~=2.28.1
Flask~=2.2.2
adapt-parser==1.0.0
lingua-franca~=0.4.3
